+++ lib/src/main/java/org/altbeacon/beacon/service/ScanJob.java
@@ -11,6 +11,7 @@ import android.content.pm.PackageItemInfo;
 import android.content.pm.PackageManager;
 import android.os.Build;
 import android.os.Handler;
+import android.os.SystemClock;
 
 import androidx.annotation.Nullable;
 
@@ -78,16 +79,17 @@ public class ScanJob extends JobService {
                 }
 
                 List<ScanResult> queuedScanResults =  new ArrayList<>(ScanJobScheduler.getInstance().dumpBackgroundScanResultQueue());
-                LogManager.d(TAG, "Processing %d queued scan resuilts", queuedScanResults.size());
+                LogManager.d(TAG, "Processing %d queued scan results", queuedScanResults.size());
                 for (ScanResult result : queuedScanResults) {
                     ScanRecord scanRecord = result.getScanRecord();
                     if (scanRecord != null) {
                         if (mScanHelper != null) {
-                            mScanHelper.processScanResult(result.getDevice(), result.getRssi(), scanRecord.getBytes());
+                            mScanHelper.processScanResult(result.getDevice(), result.getRssi(), scanRecord.getBytes(),
+                                    System.currentTimeMillis() - SystemClock.elapsedRealtime() + result.getTimestampNanos() / 1000000);
                         }
                     }
                 }
-                LogManager.d(TAG, "Done processing queued scan resuilts");
+                LogManager.d(TAG, "Done processing queued scan results");
 
                 // This syncronized block is around the scan start.
                 // Without it, it is possilbe that onStopJob is called in another thread and