+++ lib/src/test/java/org/altbeacon/beacon/AltBeaconParserTest.java
@@ -39,7 +39,7 @@ public class AltBeaconParserTest {
         BeaconManager.setDebug(true);
         byte[] bytes = hexStringToByteArray("02011a1bff1801beac2f234454cf6d4a0fadf2f4911ba9ffa600010002c50900");
         AltBeaconParser parser = new AltBeaconParser();
-        Beacon beacon = parser.fromScanData(bytes, -55, null);
+        Beacon beacon = parser.fromScanData(bytes, -55, null, 123456L);
         assertEquals ("Beacon should have one data field", 1, beacon.getDataFields().size());
         assertEquals("manData should be parsed", 9, ((AltBeacon) beacon).getMfgReserved());
     }
@@ -49,7 +49,7 @@ public class AltBeaconParserTest {
         org.robolectric.shadows.ShadowLog.stream = System.err;
         byte[] bytes = hexStringToByteArray("02011a1bff1801beac2f234454cf6d4a0fadf2f4911ba9ffa600050003be020e09526164426561636f6e20555342020a0300000000000000000000000000");
         AltBeaconParser parser = new AltBeaconParser();
-        Beacon beacon = parser.fromScanData(bytes, -55, null);
+        Beacon beacon = parser.fromScanData(bytes, -55, null, 123456L);
         assertNotNull("Beacon should be not null if parsed successfully", beacon);
     }
     @Test
@@ -58,7 +58,7 @@ public class AltBeaconParserTest {
         byte[] bytes = hexStringToByteArray("02011a1bff1801aabb2f234454cf6d4a0fadf2f4911ba9ffa600010002c50900");
         AltBeaconParser parser = new AltBeaconParser();
         parser.setMatchingBeaconTypeCode(0xaabbl);
-        Beacon beacon = parser.fromScanData(bytes, -55, null);
+        Beacon beacon = parser.fromScanData(bytes, -55, null, 123456L);
         assertNotNull("Beacon should be not null if parsed successfully", beacon);
     }
     @Test
@@ -68,7 +68,7 @@ public class AltBeaconParserTest {
         LogManager.d("XXX", "testParseWrongFormatReturnsNothing start");
         byte[] bytes = hexStringToByteArray("02011a1aff1801ffff2f234454cf6d4a0fadf2f4911ba9ffa600010002c509");
         AltBeaconParser parser = new AltBeaconParser();
-        Beacon beacon = parser.fromScanData(bytes, -55, null);
+        Beacon beacon = parser.fromScanData(bytes, -55, null, 123456L);
         LogManager.d("XXX", "testParseWrongFormatReturnsNothing end");
         assertNull("Beacon should be null if not parsed successfully", beacon);
     }
@@ -79,7 +79,7 @@ public class AltBeaconParserTest {
         org.robolectric.shadows.ShadowLog.stream = System.err;
         byte[] bytes = hexStringToByteArray("02011a1aff1801beac2f234454cf6d4a0fadf2f4911ba9ffa600010002c5000000");
         AltBeaconParser parser = new AltBeaconParser();
-        Beacon beacon = parser.fromScanData(bytes, -55, null);
+        Beacon beacon = parser.fromScanData(bytes, -55, null, 123456L);
         assertEquals("mRssi should be as passed in", -55, beacon.getRssi());
         assertEquals("uuid should be parsed", "2f234454-cf6d-4a0f-adf2-f4911ba9ffa6", beacon.getIdentifier(0).toString());
         assertEquals("id2 should be parsed", "1", beacon.getIdentifier(1).toString());