+++ lib/src/test/java/org/altbeacon/beacon/AltBeaconTest.java
@@ -55,7 +55,7 @@ public class AltBeaconTest {
     public void testRecognizeBeacon() {
         byte[] bytes = hexStringToByteArray("02011a1bff1801beac2f234454cf6d4a0fadf2f4911ba9ffa600010002c509");
         AltBeaconParser parser = new AltBeaconParser();
-        Beacon beacon = parser.fromScanData(bytes, -55, null);
+        Beacon beacon = parser.fromScanData(bytes, -55, null, 123456L);
         assertEquals("manData should be parsed", 9, ((AltBeacon) beacon).getMfgReserved() );
     }